+++ Classix Debugger/CXObjcDisassemblyWriter.mm
@@ -33,7 +33,7 @@ void CXObjCDisassemblyWriter::GetSectionMD5(const PEF::InstantiableSection &sect
 	static const char hexChars[] = {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'};
 	unsigned char md5[16];
 	CC_MD5(section.Data, section.Size(), md5);
-	for (int i = 0; i < sizeof(md5); i++)
+	for (size_t i = 0; i < sizeof(md5); i++)
 	{
 		output[i * 2] = hexChars[md5[i] >> 4];
 		output[i * 2 + 1] = hexChars[md5[i] & 0xf];
@@ -68,7 +68,7 @@ void CXObjCDisassemblyWriter::VisitOpcode(const PPCVM::Disassembly::Disassembled
 	if (!inSection) return;
 	
 	NSMutableArray* arguments = [NSMutableArray array];
-	for (int i = 0; i < opcode.Arguments.size(); i++)
+	for (size_t i = 0; i < opcode.Arguments.size(); i++)
 	{
 		NSMutableDictionary* objcOpcode = [NSMutableDictionary dictionary];
 		auto& arg = opcode.Arguments[i];
